+++ apps/templates/configuration/list-campagne.html
@@ -1,6 +1,6 @@
 {% extends "layouts/base.html" %}
 
-{% block title %} View Districts {% endblock %}
+{% block title %} Campagnes {% endblock %}
 
 {% block scriptsJs %}
 <!-- jquery -->
@@ -51,33 +51,17 @@
         <div class="container-fluid">
             <div class="row mb-2">
                 <div class="col-sm-6">
-                    <h1>Districts</h1>
+                    <h1>Campagne</h1>
                 </div>
                 <div class="col-sm-6">
                     <ol class="breadcrumb float-sm-right">
-                        <li class="breadcrumb-item"><a href="/index.html">Home</a></li>
-                        <li class="breadcrumb-item active">View Districts</li>
+                        <li class="breadcrumb-item"><a href="/index">Home</a></li>
+                        <li class="breadcrumb-item active">Campagnes</li>
                     </ol>
                 </div>
             </div>
         </div><!-- /.container-fluid -->
-        <div class="container">
-            <div class="row">
-                <div class="col-lg-3 col-6">
-                    <!-- small box -->
-                    <div class="small-box bg-success">
-                        <div class="inner">
-                            <p>Number of availabe Districts</p>
-                            <!-- <h3>53<sup style="font-size: 20px">%</sup></h3> -->
-                            <h3>{% print(num) %}</h3>
-                        </div>
-                        <div class="icon">
-                            <i class="ion ion-stats-bars"></i>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
+
         <!-- Main content -->
         <section class="content">
             <div class="container-fluid">
@@ -89,36 +73,19 @@
                     </div>
                     <div class="panel-body">
                         <div class="table-responsive">
-                            <table id="sample_data" class="text-center table table-bordered table-striped">
+                            <table id="sample_data" class="table table-bordered table-striped">
                                 <thead>
                                     <tr>
-                                        <th>Action</th>
                                         <th>Id</th>
-                                        <th>Code</th>
-                                        <th>Name</th>
-                                        <th>Action</th>
+                                        <th>code</th>
                                     </tr>
                                 </thead>
                                 <tbody>
                                     {% for c in content %}
                                     <tr>
-                                        <td>
-                                            <a class="btn btn-primary" href="../district/edit?id={{c.id}}">
-                                                <i class="fas fa-edit"></i>
-                                                Edit
-                                            </a>
-                                        </td>
                                         <td>{{c.id}}</td>
-                                        <td data-name="code" class="code" data-type="text" data-pk="{{c.id}}">
-                                            {{c.code}}</td>
-                                        <td data-name="name" class="name" data-type="text" data-pk="{{c.id}}">
+                                        <td data-name="" class="" data-type="text" data-pk="">
                                             {{c.name}}</td>
-                                        <td>
-                                            <a class="btn btn-primary" href="../district/edit?id={{c.id}}">
-                                                <i class="fas fa-edit"></i>
-                                                Edit
-                                            </a>
-                                        </td>
                                     </tr>
                                     {% endfor %}
                                 </tbody>
