+++ apps/templates/includes/sidebar.html
@@ -86,7 +86,7 @@
               </a>
             </li>
             <li class="nav-item ">
-              <a href="/configuration/groupement" class="nav-link {% if 'configuration-campagne' in segment %} active {% endif %}">
+              <a href="/configuration/campagne" class="nav-link {% if 'configuration-campagne' in segment %} active {% endif %}">
                 <i class="far fa-circle nav-icon"></i>
                 <p>Campagne</p>
               </a>
